rainbow eval runs

+++ rl-baselines/rainbow/memory.py
@@ -172,7 +172,7 @@ class ReplayMemory():
     for t in reversed(range(self.history - 1)):
       blank_mask[t] = np.logical_or(blank_mask[t + 1], transitions_firsts[t + 1]) # If future frame has timestep 0
     transitions[blank_mask] = blank_trans
-    state = torch.tensor(transitions['state'], dtype=torch.float32, device=self.device).div_(255)  # Agent will turn into batch
+    state = torch.tensor(transitions['state'].astype(np.float32), dtype=torch.float32, device=self.device).div_(255)  # Agent will turn into batch
     self.current_idx += 1
     return state
